+++ Code/mqtt_client.py
@@ -1,0 +1,42 @@
+import time
+import random
+import logging
+from paho.mqtt import client as mqtt_client
+
+_LOGGER = logging.getLogger(__name__)
+
+FIRST_RECONNECT_DELAY = 1
+RECONNECT_RATE = 2
+MAX_RECONNECT_COUNT = 12
+MAX_RECONNECT_DELAY = 60
+
+def connect_mqtt(broker, port, client_id, username, password):
+    def on_connect(client, userdata, flags, rc):
+        if rc == 0:
+            _LOGGER.info("Connected to MQTT Broker!")
+        else:
+            _LOGGER.error("Failed to connect, return code %d\n", rc)
+
+    client = mqtt_client.Client(client_id)
+    if username:
+        client.username_pw_set(username, password)
+    client.on_connect = on_connect
+    client.connect(broker, port)
+    return client
+
+def on_disconnect(client, userdata, rc):
+    _LOGGER.info("Disconnected with result code: %s", rc)
+    reconnect_count, reconnect_delay = 0, FIRST_RECONNECT_DELAY
+    while reconnect_count < MAX_RECONNECT_COUNT:
+        _LOGGER.info("Reconnecting in %d seconds...", reconnect_delay)
+        time.sleep(reconnect_delay)
+        try:
+            client.reconnect()
+            _LOGGER.info("Reconnected successfully!")
+            return
+        except Exception as err:
+            _LOGGER.error("%s. Reconnect failed. Retrying...", err)
+        reconnect_delay *= RECONNECT_RATE
+        reconnect_delay = min(reconnect_delay, MAX_RECONNECT_DELAY)
+        reconnect_count += 1
+    _LOGGER.info("Reconnect failed after %s attempts. Exiting...", reconnect_count)
